pyproject.toml
[tool.poetry]
name = "look_saver"
version = "0.1.0"
description = ""
authors = ["German <94898610+gerfest776@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.2"
djangorestframework = "^3.13.1"
psycopg2-binary = "^2.9.3"
drf-yasg = "^1.20.0"
python-decouple = "^3.6"
Pillow = "^9.0.1"
djangorestframework-simplejwt = "^5.0.0"
django-widget-tweaks = "^1.4.12"
celery = "^5.2.3"
redis = "^4.1.4"
django-filter = "^21.1"
null = "^0.6.1"
[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
